+++ NEWS.md
@@ -5,6 +5,7 @@
 * Gracefully handle non-existing module error in `box_unused_attached_mod_linter()`. (#108)
 * Support for destructure operator `%<-%`
 * Support for non-syntactic names for objects, functions, module names (#147, #151)
+* Change alphabetical sort to `radix` for cross-platform consistency. (#168)
 
 # box.linters 0.10.5